requirements.txt
streamlit
pandas
numpy
sentence-transformers
faiss-cpu
transformers
torch
scikit-learn
tqdm
wordcloud
matplotlib
